feat(dashboard): user-customizable card order + visibility
The dashboard had grown long, so let users tailor it:

- New 'Dashboard' tab in Settings (SettingsPopup is now tabbed:
  General + Dashboard). The tab lists the customizable cards as a
  vertically draggable list (native HTML5 DnD, no new deps); an eye
  toggle hides a card, moving it out of the draggable area into a
  separate 'Hidden cards' list. Reset restores defaults.
- Dashboard renders the customizable region in a single column from the
  persisted order, excluding hidden cards. The top KPI header
  (net worth / bank / investments / cash) stays pinned and is not part
  of the customizable set.
- useDashboardLayout hook persists order+hidden to localStorage with a
  pub/sub + stable cached snapshot (mirrors useBudgetAlertSettings);
  normalizes on read so cards added in future releases surface
  automatically rather than being dropped.
- Extracted the large tabbed analytics block into DashboardChartsPanel
  (self-contained, queries deduped by React Query key) so it can be one
  orderable/hideable card.
- Added a Settings gear to the mobile top bar next to the alerts bell.
- en/he locale keys for the new tab, card labels, and controls.
- e2e spec covering hide / persistence+restore / pinned header; existing
  dashboard specs still green.
