+++ frontend/src/locales/en.json
@@ -103,7 +103,16 @@
     "backupCreated": "Backup created successfully",
     "backupRestored": "Backup restored successfully. Refreshing...",
     "backupFailed": "Backup failed",
-    "restoreFailed": "Restore failed"
+    "restoreFailed": "Restore failed",
+    "tabGeneral": "General",
+    "tabDashboard": "Dashboard",
+    "dashboardLayoutHelp": "Drag to reorder cards. Hidden cards are removed from the dashboard.",
+    "dashboardLayoutReset": "Reset",
+    "dashboardVisibleCards": "Visible cards",
+    "dashboardHiddenCards": "Hidden cards",
+    "dashboardNoVisibleCards": "All cards are hidden",
+    "dashboardHideCard": "Hide card",
+    "dashboardShowCard": "Show card"
   },
   "dashboard": {
     "title": "Dashboard",
@@ -240,6 +249,16 @@
     "heatmap": {
       "title": "Spending Calendar",
       "monthTotal": "This month"
+    },
+    "cards": {
+      "forecast": "This Month (forecast)",
+      "insights": "Insights",
+      "budget": "Budget spending",
+      "recent": "Recent transactions",
+      "recurring": "Subscriptions & recurring",
+      "goals": "Savings goals",
+      "heatmap": "Spending calendar",
+      "charts": "Charts & analytics"
     }
   },
   "transactions": {
